Change to Long the primary key datatype

+++ src/main/java/co/org/osso/wallet/AccountRepository.java
@@ -2,5 +2,5 @@ package co.org.osso.wallet;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Integer> {
+public interface AccountRepository extends CrudRepository<Account, Long> {
 }
